requirements.txt
anthropic
instructor
pydantic
PyPDF2
pdf2image
Pillow